requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
flask==3.1.0
python-dotenv==1.1.1
pytz==2023.3
requests==2.31.0
